! 2012-02-21 - Chris Jones  - Update the way some references to Particles and Vertices are stored, to    use 64 bit intergers, to avoid packing problems.
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@135640 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 61466d4a07ab8b4390b91e0af85737738787ede7

+++ LHCb/Event/EventPacker/src/ReadPackedDst.cpp
@@ -263,7 +263,7 @@ StatusCode ReadPackedDst::execute() {
       processLinks( parts, version );
       getFromBlob<LHCb::PackedParticle> ( parts->data(), blobs );
       getFromBlob<std::pair<int,int> >  ( parts->extra(), blobs );
-      getFromBlob<int>                  ( parts->daughters(), blobs );
+      getFromBlob<long long>            ( parts->daughters(), blobs );
 
     } else if ( LHCb::CLID_PackedVertices        == classID ) {
 
@@ -271,7 +271,7 @@ StatusCode ReadPackedDst::execute() {
       put( verts, name + m_postfix );
       processLinks( verts, version );
       getFromBlob<LHCb::PackedVertex> ( verts->data(), blobs );
-      getFromBlob<int>                ( verts->outgoingParticles(), blobs );
+      getFromBlob<long long>          ( verts->outgoingParticles(), blobs );
 
     } else if ( LHCb::CLID_PackedWeightsVector   == classID ) {
